move VERIFICATION_TYPES to core
