+++ packages/tenderly-core/src/internal/core/types/index.ts
@@ -6,3 +6,4 @@ export * from "./Responses";
 export * from "./TenderlyNetwork";
 export * from "./VerifyContractABIRequest";
 export * from "./VerifyContractABIResponse";
+export * from "./verification-types";
