+++ packages/tenderly-core/src/internal/core/types/verification-types.ts
@@ -1,0 +1,6 @@
+export const VERIFICATION_TYPES = {
+  FORK: "fork",
+  PRIVATE: "private",
+  PUBLIC: "public",
+  DEVNET: "devnet",
+};
