+++ packages/tenderly-hardhat/src/constants.ts
@@ -1,10 +1,4 @@
 export const PLUGIN_NAME = "hardhat-tenderly";
 export const DEFAULT_CHAIN_ID = "31337";
-export const VERIFICATION_TYPES = {
-  FORK: "fork",
-  PRIVATE: "private",
-  PUBLIC: "public",
-  DEVNET: "devnet",
-};
 export const CONTRACT_NAME_PLACEHOLDER =
   "HARDHAT_TENDERLY_CONTRACT_NAME_PLACEHOLDER";
